pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "consilience"
version = "0.1.0"
description = "Production-grade multi-agent research orchestration platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.15",
    "anthropic>=0.12.0",
    "openai>=1.12.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "stripe>=5.0.0",
    "bcrypt>=4.0.0",
    "pyjwt>=2.8.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "duckduckgo-search>=5.0.0",
    "tavily-python>=0.3.0",
    "python-dateutil>=2.8.2",
    "click>=8.1.7",
    "rich>=13.7.0",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.0.0",
    "python-jose>=3.3.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.15.0",
    "black>=23.11.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]  
include = [
    "api*",
    "core*",
    "models*",
    "agents*",
    "config*",
    "services*",
    "database*",
    "orchestrator*",
    "tools*",
    "utils*",
]
